Extract RequestBuilder
